feat: Add rock-ci-metadata.yaml (#21)
* chore(refactor): Move rockcraft files to separate folder

As we have agreed in the past (see
canonical/charmed-kubeflow-workflows#52), single rock repositories
should also contain a directory for the rock. Thus, move all rockcraft
specific files to a folder of the same name as the rock.

* feat: Add rock-ci-metadata.yaml

Introduce a `rock-ci-metadata.yaml` for rock integration. For full file
creation steps, refer to
https://github.com/canonical/pipelines-rocks/issues/200#issuecomment-3112245114.

Closes #19